+++ xclim/testing/tests/test_sdba/test_base.py
@@ -1,3 +1,4 @@
+import jsonpickle
 import numpy as np
 import pytest
 import xarray as xr
@@ -23,11 +24,9 @@ def test_param_class():
         "group=Grouper(dim='time',"
     )
 
-    s = str(obj)
-    obj2 = Parametrizable.from_string(s)
+    s = jsonpickle.encode(obj)
+    obj2 = jsonpickle.decode(s)
     assert obj.parameters == obj2.parameters
-
-    assert "ATestSubClass" in Parametrizable._get_subclasses()
 
 
 @pytest.mark.parametrize(